+++ i18n/locales/en/common.json
@@ -444,6 +444,7 @@
   "Sign in with a hardware wallet": "Sign in with a hardware wallet",
   "Sign message": "Sign message",
   "Sign multisignature transaction": "Sign multisignature transaction",
+  "Sign out": "Sign out",
   "Sign out automatically after 10 minutes.": "Sign out automatically after 10 minutes.",
   "Signature": "Signature",
   "Signed": "Signed",
