+++ src/components/shared/navigationBars/topBar/topBar.css
@@ -141,10 +141,8 @@
 }
 
 .logoutBtn {
-  position: relative;
-  display: inline-block;
-  vertical-align: middle;
   display: flex;
   align-items: center;
+  height: 100%;
   padding-right: 24px;
 }
